// src/General/interfaces/pv_estados.ts
export interface Ipv_estados {
  id: number;
  momento: string;
  tipo: number;
  eco: number;
  eco_estatus: number;
  eco_tipo: number;
  motivo_id: number;
  motivo_desc: string;
  modulo: number;
  direccion: string;
  ruta: string;
  ruta_modalidad: string;
  ruta_cc: string;
  op_cred: number;
  op_turno: number;
  extintor: string;
}